% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellassign.R
\name{cellassign}
\alias{cellassign}
\title{Assign each square to a cell}
\usage{
cellassign(visiumpath, segmentationpath, cellnucratio = 2)
}
\arguments{
\item{visiumpath}{The path to the Visium HD output folder}

\item{segmentationpath}{The path to the StarDist segmentation file}

\item{cellnucratio}{The ratio of cell size and nucleus size (2 by default)}
}
\value{
A list including a vector indicating which cell each square belongs to, and the matrix of spatial locations
}
\description{
Assign each square to a cell
}
