% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsvg.R
\name{ctsvg_fit}
\alias{ctsvg_fit}
\title{Fit spatial gene expression patterns within each cell cluster}
\usage{
ctsvg_fit(d, coord, cell.filter = TRUE)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{coord}{The matrix of spatial locations from \code{\link{cellassign}}}

\item{cell.filter}{Whether to remove spatially isolated cells within each cell cluster (\code{TRUE} by default)}
}
\value{
The fitted cell-cluster-specific gene expression matrix
}
\description{
Fit spatial gene expression patterns within each cell cluster
}
