% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmodule.R
\name{ctsvg_module}
\alias{ctsvg_module}
\title{Identify gene modules within each cell cluster}
\usage{
ctsvg_module(d, clutestRes, clufitRes, seed = 2024, clunum = NULL)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{clutestRes}{The output from \code{\link{ctsvg_test}}}

\item{clufitRes}{The output from \code{\link{ctsvg_fit}}}

\item{seed}{Random seed for k-means clustering (2024 by default)}

\item{clunum}{Number of modules (\code{NULL} by default, automatically selected)}
}
\value{
A list with each element representing a cell cluster and indicating which module each gene belongs to
}
\description{
Identify gene modules within each cell cluster
}
