% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsvg.R
\name{ctsvg_test}
\alias{ctsvg_test}
\title{Identify spatially variable genes within each cell cluster}
\usage{
ctsvg_test(
  d,
  coord,
  cell.filter = TRUE,
  clu.list = NULL,
  jac.mat = NULL,
  min.cells = 0.01,
  seed = 2024,
  verbose = FALSE,
  n.permute = 100
)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{coord}{The matrix of spatial locations from \code{\link{cellassign}}}

\item{cell.filter}{Whether to remove spatially isolated cells within each cell cluster (\code{TRUE} by default)}

\item{clu.list}{The list of reassigned clusters from \code{\link{recluster}}}

\item{jac.mat}{The matrix of Jaccard indices from \code{\link{recluster}}}

\item{min.cells}{The minimum proportion of cells within each cell cluster a gene is expressed in to be retained (0.01 by default)}

\item{seed}{Random seed for permutation (2024 by default)}

\item{verbose}{Whether to print the cell cluster in progress (\code{FALSE} by default)}

\item{n.permute}{Number of permutation times (reassigned clusters with the highest Jaccard indices, 100 by default)}
}
\value{
A data frame with p-values and FDRs for each gene in each cell cluster
}
\description{
Identify spatially variable genes within each cell cluster
}
