% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexpr.R
\name{getexpr}
\alias{getexpr}
\title{Generate single-cell gene expression profiles}
\usage{
getexpr(cellassignRes, visiumpath)
}
\arguments{
\item{cellassignRes}{The output from \code{\link{cellassign}}}

\item{visiumpath}{The path to the Visium HD output folder}
}
\value{
The single-cell gene expression count matrix
}
\description{
Generate single-cell gene expression profiles
}
