% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotGene}
\alias{plotGene}
\title{Visualize spatial patterns of a single gene}
\usage{
plotGene(
  d,
  clufitRes = NULL,
  coord,
  gene,
  cellclu = NULL,
  background = FALSE,
  ...
)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{clufitRes}{The output from \code{\link{ctsvg_fit}} (\code{NULL} by default, original gene expression matrix)}

\item{coord}{The matrix of spatial locations from \code{\link{cellassign}}}

\item{gene}{Which gene to plot}

\item{cellclu}{A character value indicating which cell cluster to plot (\code{NULL} by default, all cells)}

\item{background}{Whether to plot all cells (\code{FALSE} by default)}

\item{...}{The parameters in the \code{geom_point(...)} function}
}
\value{
A ggplot2 plot
}
\description{
Visualize spatial patterns of a single gene
}
