% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMetagene}
\alias{plotMetagene}
\title{Visualize spatial patterns of metagenes within a cell cluster}
\usage{
plotMetagene(d, clufitRes, moduleRes, coord, cluster, background = FALSE, ...)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{clufitRes}{The output from \code{\link{ctsvg_fit}}}

\item{moduleRes}{The output from \code{\link{ctsvg_module}}}

\item{coord}{The matrix of spatial locations from \code{\link{cellassign}}}

\item{cluster}{A character value indicating which cell cluster to plot}

\item{background}{Whether to plot all cells (\code{FALSE} by default)}

\item{...}{The parameters in the \code{geom_point(...)} function}
}
\value{
A ggplot2 plot
}
\description{
Visualize spatial patterns of metagenes within a cell cluster
}
