% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster.R
\name{recluster}
\alias{recluster}
\title{Reassign cells to clusters}
\usage{
recluster(d, n.max = 1000, resolution = 1.2)
}
\arguments{
\item{d}{The output from \code{\link{runseurat}}}

\item{n.max}{Number of reclustering times}

\item{resolution}{The parameter used in \code{\link{runseurat}} (1.2 by default)}
}
\value{
A list including the matrix of Jaccard indices, and the list of reassigned clusters
}
\description{
Reassign cells to clusters
}
