% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runseurat.R
\name{runseurat}
\alias{runseurat}
\title{Run Seurat pipeline}
\usage{
runseurat(
  count,
  min.genes = 300,
  min.cells = 0.01,
  npcs = 10,
  resolution = 1.2
)
}
\arguments{
\item{count}{The output from \code{\link{getexpr}}}

\item{min.genes}{The minimum number of genes that are expressed in a cell for it to be retained (300 by default)}

\item{min.cells}{The minimum proportion of cells a gene is expressed in to be retained (0.01 by default)}

\item{npcs}{Number of PCs to use (10 by default)}

\item{resolution}{The parameter in the \code{FindClusters} function (1.2 by default)}
}
\value{
A Seurat object with cell clustering results
}
\description{
Run Seurat pipeline
}
